2022-11: Don't use modular arithmetic when division is involved.
Earlier, the code computed the worry level for both parts like so:
    worry = (ops[m](item) // divisor) % LCM

For Part 2, this is fine, since the divisor = 1 (i.e., nothing is
actually getting divided).

For Part 1 however, the divisor = 3, so it's wrong. Division is NOT
compatible under modular arithmetic, except in specific circumstances.
References:
    https://en.wikipedia.org/wiki/Modular_arithmetic#Properties
    https://www.doc.ic.ac.uk/~mrh/330tutor/ch03.html

I realized this as I was writing comments on why modular arithmetic is
provably correct for Part 2. I couldn't prove it for Part 1!

I hadn't discovered the issue earlier, because the wrong code above
still produces the correct answer for the AoC inputs (both example and
real inputs), by sheer luck. The LCM in those inputs is so large, and
"x // 3" is always small enough that the "% LCM" is a no-op.

That gives us a hint on how to construct a counter-example:
    Monkey 0:
      Starting items: 209
      Operation: new = old + 1
      Test: divisible by 2
        If true: throw to monkey 1
        If false: throw to monkey 2

    Monkey 1:
      Starting items: 1
      Operation: new = old + 1
      Test: divisible by 5
        If true: throw to monkey 2
        If false: throw to monkey 0

    Monkey 2:
      Starting items: 1
      Operation: new = old + 1
      Test: divisible by 7
        If true: throw to monkey 0
        If false: throw to monkey 1

In the above example, LCM = 2 * 5 * 7 = 70. When Monkey 0 processes the
first item (209), consider the two versions of the code:
    Right: worry = ((209 + 1) // 3)       = 70
    Wrong: worry = ((209 + 1) // 3) % 70  =  0
Regardless, the divisibility test gives the same result (both 70 and 0
are divisible by 2), so the item is thrown to Monkey 1. But then see
what happens when Monkey 1 processes the item:
    Right: worry = ((70 + 1) // 3)       = 23
    Wrong: worry = (( 0 + 1) // 3) % 70  =  0
Now the divisibility test gives different results!